Changed class into entity
